Initial commit: Set up Kotlin project with Gradle
This commit creates the initial setup for a Kotlin project with Gradle. This setup includes the source code files, Gradle build files, and configuration files for the IntelliJ IDEA project setup. Also, Kotlin coding style setting is established. All these settings form the basis on which we build our project. This setup operation is necessary to kick off the project.

Signed-off-by: Arnau Mora <arnyminerz@proton.me>